requirements.txt
chromedriver-binary==88.0.4324.96.0
chromedriver-binary-auto==0.1
click==7.1.2
cycler==0.10.0
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
japanize-matplotlib==1.1.3
Jinja2==2.11.2
kiwisolver==1.3.1
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.4
numpy==1.20.1
pandas==1.2.1
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2021.1
scipy==1.6.0
seaborn==0.11.1
selenium==3.141.0
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
XlsxWriter==1.3.7
